replacing standard ruby Enumerator with a NilEnumerator
the code suffered tremendously due to the fact that the standard enumerator throws an exception if you try to do next/peek when it has reached the end. By replacing it with an enumerator that just returns nil we can simplify the code by removing the intermediate array and operations on that array.